Implement WEBTOYS.ai redesign and fix TypeScript build error
- Create new WEBTOYS.ai landing page with playful, colorful design
- Move original cyberpunk WTAF design to /wtaf-landing-old backup
- Replace main /wtaf-landing with new WEBTOYS.ai design
- Add interactive features: SMS bubbles, remix functionality, Konami code
- Update middleware to handle backup URL routing
- Fix TypeScript error in ZAD save route with proper null checking
- Maintain full responsiveness and existing SMS integration
