+++ web/app/api/zad/save/route.ts
@@ -444,6 +444,10 @@ export async function POST(req: NextRequest) {
           n: 1
         });
 
+        if (!response.data || response.data.length === 0) {
+          throw new Error("No image data in DALL-E response");
+        }
+        
         const imageUrl = response.data[0].url;
         if (!imageUrl) {
           throw new Error("No image URL in DALL-E response");
